# openapi/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: go-housework
  description: 家事タスクを管理するAPIサーバです。
servers:
  - url: http://localhost:8080
    description: Development local server
  - url: https://XXXX.com
    description: Production server
paths:
...
  /family:
    post:
      security:
        - bearerAuth: []
      summary: 世帯登録API
      description: 世帯を登録するAPI。登録者が世帯管理者となる。世帯管理者は変更できない。
      operationId: createFamily
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestCreateFamily'
      responses:
        '200':
          description: Success Response
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ResponseCreateFamily'
        '400':
          description: Bad Request Response
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ResponseBadRequest'
        '401':
          description: Unauthorized
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ResponseUnauthorized'
        '500':
          description: Internal Server Error
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ResponseInternalServerError'
    get:
      security:
        - bearerAuth: []
      summary: 世帯詳細情報取得API
      description: 世帯詳細情報を取得するAPI。世帯管理者のみ実行可能。
      operationId: showFamily
      responses:
        '200':
          description: Success Response
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ResponseShowFamily'
        '401':
          description: Unauthorized
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ResponseUnauthorized'
        '500':
          description: Internal Server Error
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ResponseInternalServerError'
    put:
      security:
        - bearerAuth: []
      summary: 世帯情報更新API
      description: 世帯情報を更新するAPI。世帯管理者のみ実行可能。
      operationId: updateFamily
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestUpdateFamily'
      responses:
        '200':
          description: Success Response
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ResponseUpdateFamily'
        '400':
          description: Bad Request Response
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ResponseBadRequest'
        '401':
          description: Unauthorized
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ResponseUnauthorized'
        '500':
          description: Internal Server Error
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ResponseInternalServerError'
    delete:
      security:
        - bearerAuth: []
      summary: 世帯情報削除API
      description: 世帯情報を削除するAPI。世帯管理者のみ実行可能。
      operationId: deleteFamily
      responses:
        '200':
          description: Success Response
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ResponseDeleteFamily'
        '401':
          description: Unauthorized
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ResponseUnauthorized'
        '500':
          description: Internal Server Error
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ResponseInternalServerError'
...
  /task:
    post:
      security:
        - bearerAuth: []
      summary: 家事タスク登録API
      description: 自身の所属する世帯に家事タスクを作成するAPI
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestCreateTask'
      responses:
        '200':
          description: Success Response
          content:

            'application/json':
              schema:
                $ref: '#/components/schemas/ResponseCreateTask'
        '400':
          description: Bad Request Response
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ResponseBadRequest'
        '401':
          description: Unauthorized
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ResponseUnauthorized'
        '500':
          description: Internal Server Error
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ResponseInternalServerError'
    put:
      security:
        - bearerAuth: []
      summary: 家事タスク更新API
      description: 自身の所属する世帯の家事タスクを更新するAPI。タスク名、担当者名、ステータスを変更可能。
      operationId: updateTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestUpdateTask'
      responses:
        '200':
          description: Success Response
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ResponseUpdateTask'
        '400':
          description: Bad Request Response
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ResponseBadRequest'
        '401':
          description: Unauthorized
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ResponseUnauthorized'
        '500':
          description: Internal Server Error
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ResponseInternalServerError'
...
  /tasks:
    get:
      security:
        - bearerAuth: []
      summary: 家事タスク一覧取得API
      description: 自身の所属する指定日付の世帯の家事タスク一覧を取得するAPI
      operationId: listTasks
      parameters:
        - in: query
          name: date
          description: 担当日付
          required: true
          schema:
            type: string
            example: "2020-03-01"
      responses:
        '200':
          description: Success Response
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ResponseListTasks'
        '400':
          description: Bad Request Response
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ResponseBadRequest'
        '401':
          description: Unauthorized
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ResponseUnauthorized'
        '500':
          description: Internal Server Error
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ResponseInternalServerError'
components:
  securitySchemes:
    bearerAuth:
      description: JWT
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Family:
      description: Familyスキーマ
      type: object
      properties:
        family_id:
          description: 家族ID
          type: integer
          format: int64
          example: 1
        family_name:
          description: 家族名
          type: string
          example: 'James'
      required:
        - family_id
        - family_name
...
    Task:
      description: Taskスキーマ
      type: object
      properties:
        task_id:
          description: 家事タスクID
          type: integer
          format: int64
          example: 1
        task_name:
          description: 家事タスク名
          type: string
          example: 'Cooking lunch'
        member_name:
          description: ユーザ名。家事の担当者。
          type: string
          example: 'LeBron'
        status:
          description: 家事進捗ステータス
          type: string
          enum:
            - 'todo'
            - 'done'
        date:
          description: 担当日付
          type: string
          example: "2020-03-01"
      required:
        - task_id
        - task_name
        - member_name
        - status
        - date
    # Request
...
    RequestCreateFamily:
      description: createFamilyのリクエストスキーマ
      type: object
      properties:
        family_name:
          description: 家族名
          type: string
          example: 'James'
      required:
        - family_name
    RequestUpdateFamily:
      description: updateFamilyのリクエストスキーマ。現状、家族名のみ変更可能なため、family_nameを必須プロパティとしている。
      type: object
      properties:
        family_name:
          description: 家族名
          type: string
          example: 'James'
      required:
        - family_name
...
    RequestCreateTask:
      description: createTaskのリクエストスキーマ
      type: object
      properties:
        task_name:
          description: 家事タスク名
          type: string
          example: 'Cooking lunch'
        member_name:
          description: ユーザ名。家事の担当者。
          type: string
          example: 'LeBron'
        status:
          description: 家事進捗ステータス
          type: string
          enum:
            - 'todo'
            - 'done'
        date:
          description: 担当日付
          type: string
          example: "2020-03-01"
      required:
        - task_name
        - member_name
        - status
        - date
    RequestUpdateTask:
      description: updateTaskのリクエストスキーマ
      type: object
      properties:
        task:
          $ref: '#/components/schemas/Task'
      required:
        - task

    # Response
...
    ResponseCreateFamily:
      description: 世帯登録APIの正常系レスポンススキーマ
      type: object
      properties:
        family:
          $ref: '#/components/schemas/Family'
      required:
        - family
    ResponseUpdateFamily:
      description: 世帯更新APIの正常系レスポンススキーマ
      type: object
      properties:
        family:
          $ref: '#/components/schemas/Family'
      required:
        - family
    ResponseShowFamily:
      description: 世帯詳細情報取得APIの正常系レスポンススキーマ
      type: object
      properties:
        family:
          $ref: '#/components/schemas/Family'
      required:
        - family
    ResponseDeleteFamily:
      description: 世帯削除APIの正常系レスポンススキーマ
      type: object
      properties:
        message:
          description: HTTPレスポンスメッセージ
          type: string
          example: 'the family deleted'
      required:
        - message
...
    ResponseCreateTask:
      description: 家事タスク作成APIの正常系レスポンススキーマ
      type: object
      properties:
        family:
          $ref: '#/components/schemas/Family'
        task:
          $ref: '#/components/schemas/Task'
      required:
        - family
        - task
    ResponseUpdateTask:
      description: 家事タスク更新APIの正常系レスポンススキーマ
      type: object
      properties:
        family:
          $ref: '#/components/schemas/Family'
        task:
          $ref: '#/components/schemas/Task'
      required:
        - family
        - task
...
    ResponseListTasks:
      description: 家事タスク一覧取得APIの正常系レスポンススキーマ
      type: object
      properties:
        family:
          $ref: '#/components/schemas/Family'
        tasks:
          description: 家事タスク情報の集合
          type: array
          items:
            $ref: '#/components/schemas/Task'
      required:
        - family
        - tasks
    ResponseBadRequest:
      description: Bad Reqeust
      type: object
      properties:
        message:
          description: HTTPレスポンスメッセージ
          type: string
          example: 'Something Bad Request Error'
      required:
        - message
    ResponseUnauthorized:
      description: Unauthorized
      type: object
      properties:
        message:
          description: HTTPレスポンスメッセージ
          type: string
          example: 'Something Unauthorized Error'
      required:
        - message
...
    ResponseInternalServerError:
      description: Internal Server Error
      type: object
      properties:
        message:
          description: HTTPレスポンスメッセージ
          type: string
          example: 'Something Internal Server Error'
      required:
        - message